+++ CHANGES.rst
@@ -73,6 +73,7 @@ Pull requests
 - [#395] Fix misc issues with IRF classes (Johannes King)
 - [#394] Move some data specs to gamma-astro-data-formats (Christoph Deil)
 - [#392] Use external ci-helpers (Brigitta Sipocz)
+- [#387] Improve Gammapy catalog query and browser (Christoph Deil)
 - [#379] Add gammapy.region and reflected region computation (Johannes King)
 - [#375] Misc cleanup of scripts and docs (Christoph Deil)
 - [#371] Improve catalog utils (Christoph Deil)